+++ src/async.ts
@@ -1,1 +1,0 @@
-export type SyncOrAsync<T> = T | Promise<T>;