+++ src/handle-permission-request.ts
@@ -1,5 +1,0 @@
-import { SyncOrAsync } from "./async.js";
-
-export type HandlePermissionRequest = (
-  descriptor: PermissionDescriptor,
-) => SyncOrAsync<PermissionState>;