+++ test/vitest/user.spec.ts
@@ -1,5 +1,5 @@
 import {
-  HandlePermissionRequest,
+  HandleAccessRequest,
   PermissionStore,
   User,
   createPermissionStore,
@@ -31,82 +31,157 @@ describe("User", () => {
     });
   });
 
-  describe("when no permission request callback is configured", () => {
+  describe("when no access request handler is configured", () => {
     beforeEach(() => {
       user = createUser({ permissionStore });
     });
 
-    describe('when a permission in the "prompt" state is requested', () => {
-      it("denies the permission", async () => {
-        expect(await user.requestPermission(permissionA)).toBe("denied");
-        expect(permissionStore.get(permissionA)).toBe("denied");
+    describe('when access is requested for a permission in the "prompt" state', () => {
+      it("denies access and leaves the permission unchanged", async () => {
+        expect(await user.requestAccess(permissionA)).toBe(false);
+        expect(permissionStore.get(permissionA)).toBe("prompt");
       });
     });
 
-    describe('when a permission in the "granted" state is requested', () => {
-      it("leaves the permission granted", async () => {
-        expect(await user.requestPermission(permissionB)).toBe("granted");
+    describe('when access is requested for a permission in the "granted" state', () => {
+      it("allows access and leaves the permission unchanged", async () => {
+        expect(await user.requestAccess(permissionB)).toBe(true);
         expect(permissionStore.get(permissionB)).toBe("granted");
       });
     });
 
-    describe('when a permission in the "denied" state is requested', () => {
-      it("leaves the permission denied", async () => {
-        expect(await user.requestPermission(permissionC)).toBe("denied");
+    describe('when access is requested for a permission in the "denied" state', () => {
+      it("denies access and leaves the permission unchanged", async () => {
+        expect(await user.requestAccess(permissionC)).toBe(false);
         expect(permissionStore.get(permissionC)).toBe("denied");
       });
     });
   });
 
-  describe("when a permission request callback is configured", () => {
-    let handlePermissionRequest: Mock<HandlePermissionRequest>;
+  describe("when an access request handler is configured", () => {
+    let handleAccessRequest: Mock<HandleAccessRequest>;
 
     beforeEach(() => {
-      handlePermissionRequest = vi.fn(
-        // eslint-disable-next-line @typescript-eslint/no-misused-promises
-        (async () => "granted") as HandlePermissionRequest,
-      );
+      handleAccessRequest = vi.fn(async (dialog) => {
+        dialog.allow(true);
+      });
 
-      user = createUser({ permissionStore, handlePermissionRequest });
+      user = createUser({ permissionStore, handleAccessRequest });
     });
 
-    describe('when a permission in the "prompt" state is requested', () => {
-      it("calls the callback with the permission descriptor", async () => {
-        await user.requestPermission(permissionA);
+    it("can't dismiss the dialog after it's been dismissed", async () => {
+      handleAccessRequest.mockImplementation(async (dialog) => {
+        dialog.allow(true);
+        dialog.deny(true);
+      });
 
-        expect(handlePermissionRequest).toHaveBeenCalledWith({
-          name: "permission-a",
+      await expect(user.requestAccess(permissionA)).rejects.toThrow(
+        "Access dialog already dismissed",
+      );
+    });
+
+    describe('when access is requested for a permission in the "prompt" state', () => {
+      it("calls the callback with a dialog and the permission descriptor", async () => {
+        await user.requestAccess(permissionA);
+
+        expect(handleAccessRequest).toHaveBeenCalledWith(
+          expect.objectContaining({
+            dismiss: expect.any(Function) as () => void,
+            allow: expect.any(Function) as () => void,
+            deny: expect.any(Function) as () => void,
+          }),
+          { name: "permission-a" },
+        );
+      });
+
+      describe("when the dialog is dismissed", () => {
+        beforeEach(() => {
+          handleAccessRequest.mockImplementation(async (dialog) => {
+            dialog.dismiss();
+          });
+        });
+
+        it("denies access and leaves the permission unchanged", async () => {
+          expect(await user.requestAccess(permissionA)).toBe(false);
+          expect(permissionStore.get(permissionA)).toBe("prompt");
         });
       });
 
-      it("updates the permission state with the callback's return value", async () => {
-        expect(await user.requestPermission(permissionA)).toBe("granted");
-        expect(permissionStore.get(permissionA)).toBe("granted");
+      describe("when access is allowed but the decision is not remembered", () => {
+        beforeEach(() => {
+          handleAccessRequest.mockImplementation(async (dialog) => {
+            dialog.allow(false);
+          });
+        });
+
+        it("allows access and leaves the permission unchanged", async () => {
+          expect(await user.requestAccess(permissionA)).toBe(true);
+          expect(permissionStore.get(permissionA)).toBe("prompt");
+        });
+      });
+
+      describe("when access is allowed and the decision is remembered", () => {
+        beforeEach(() => {
+          handleAccessRequest.mockImplementation(async (dialog) => {
+            dialog.allow(true);
+          });
+        });
+
+        it("allows access and grants the permission", async () => {
+          expect(await user.requestAccess(permissionA)).toBe(true);
+          expect(permissionStore.get(permissionA)).toBe("granted");
+        });
+      });
+
+      describe("when access is denied but the decision is not remembered", () => {
+        beforeEach(() => {
+          handleAccessRequest.mockImplementation(async (dialog) => {
+            dialog.deny(false);
+          });
+        });
+
+        it("denies access and leaves the permission unchanged", async () => {
+          expect(await user.requestAccess(permissionA)).toBe(false);
+          expect(permissionStore.get(permissionA)).toBe("prompt");
+        });
+      });
+
+      describe("when access is denied and the decision is remembered", () => {
+        beforeEach(() => {
+          handleAccessRequest.mockImplementation(async (dialog) => {
+            dialog.deny(true);
+          });
+        });
+
+        it("denies access and denies the permission", async () => {
+          expect(await user.requestAccess(permissionA)).toBe(false);
+          expect(permissionStore.get(permissionA)).toBe("denied");
+        });
       });
     });
 
-    describe('when a permission in the "granted" state is requested', () => {
+    describe('when access is requested for a permission in the "granted" state', () => {
       it("does not call the callback", async () => {
-        await user.requestPermission(permissionB);
+        await user.requestAccess(permissionB);
 
-        expect(handlePermissionRequest).not.toHaveBeenCalled();
+        expect(handleAccessRequest).not.toHaveBeenCalled();
       });
 
-      it("leaves the permission granted", async () => {
-        expect(await user.requestPermission(permissionB)).toBe("granted");
+      it("allows access and leaves the permission unchanged", async () => {
+        expect(await user.requestAccess(permissionB)).toBe(true);
         expect(permissionStore.get(permissionB)).toBe("granted");
       });
     });
 
-    describe('when a permission in the "denied" state is requested', () => {
+    describe('when access is requested for a permission in the "denied" state', () => {
       it("does not call the callback", async () => {
-        await user.requestPermission(permissionC);
+        await user.requestAccess(permissionC);
 
-        expect(handlePermissionRequest).not.toHaveBeenCalled();
+        expect(handleAccessRequest).not.toHaveBeenCalled();
       });
 
-      it("leaves the permission denied", async () => {
-        expect(await user.requestPermission(permissionC)).toBe("denied");
+      it("denies access and leaves the permission unchanged", async () => {
+        expect(await user.requestAccess(permissionC)).toBe(false);
         expect(permissionStore.get(permissionC)).toBe("denied");
       });
     });